Add StickerDecorator
